worker/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "speakflow-worker"
version = "1.0.0"
description = "SpeakFlow Worker - Audio analysis pipeline"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "redis>=5.0.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "boto3>=1.34.0",
    "openai-whisper>=20231117",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "openai>=1.10.0",
    "httpx>=0.26.0",
    "speakflow-contracts",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
